cloud_run_code/requirements.txt
Flask==1.1.2
google-api-python-client
google-auth==1.28.0
google-cloud-pubsub
google-cloud-storage
gunicorn
